Forgot Cobi's
git-svn-id: https://waca.svn.sourceforge.net/svnroot/waca/trunk/acc@381 41b7b2fa-e25d-4c8f-a56d-f93ce3f2ce94

+++ accbot.php
@@ -282,7 +282,7 @@ while (!feof($fp)) {
 		$nick = substr($nick[0],1);
 		$hostA = explode('@',$line_ex[0]);
 		$host = $hostA[1];
-		if (($nick == 'Cobi') or ($nick == 'SQLDb') && ($host = 'wikipedia/SQL') or ($nick == '|Cobi|') or ($nick == 'Cobi-Laptop')) {
+		if (($nick == 'Cobi') && ($host = 'cobi.cluenet.org') or ($nick == 'SQLDb') && ($host = 'wikipedia/SQL') or ($nick == '|Cobi|') or ($nick == 'Cobi-Laptop')) {
 			fwrite($fp,'PRIVMSG '.$chan.' :'.$nick.": /whois $nick\n");
 			if (pcntl_fork() == 0) {
 				$svn = popen('svn up 2>&1', 'r');
@@ -323,7 +323,7 @@ while (!feof($fp)) {
                 $nick = substr($nick[0],1);
 
 
-                if (($nick == 'Cobi') or ($nick == 'SQLDb') && ($host = 'wikipedia/SQL') or ($nick == '|Cobi|') or ($nick == 'Cobi-Laptop')) {
+                if (($nick == 'Cobi') && ($host = 'cobi.cluenet.org') or ($nick == 'SQLDb') && ($host = 'wikipedia/SQL') or ($nick == '|Cobi|') or ($nick == 'Cobi-Laptop')) {
 			echo 'Restart from IRC!';
 			fclose($fp);
 			fclose($fpt);
@@ -335,7 +335,7 @@ while (!feof($fp)) {
 		$nick = explode('!',$line_ex[0]);
 		$nick = substr($nick[0],1);
 
-		 if (($nick == 'Cobi') or ($nick == 'SQLDb') && ($host = 'wikipedia/SQL') or ($nick == '|Cobi|') or ($nick == 'Cobi-Laptop')) {
+		 if (($nick == 'Cobi') && ($host = 'cobi.cluenet.org') or ($nick == 'SQLDb') && ($host = 'wikipedia/SQL') or ($nick == '|Cobi|') or ($nick == 'Cobi-Laptop')) {
 			fwrite($fp,'PRIVMSG '.$chan.' :'.$nick.': Please wait while I try to fix the SVN.'."\n");
 			system('tar -jcvpf ~/accinterface-svn-broken.'.time().'.tbz2 .');
 			system('svn list | xargs rm -f');
